Add empty /etc/pacman.d/mirrorlist file.
pacstrap expects /etc/pacman.d/mirrorlist with a valid mirror.
It copies the file to the new root and shows an error if it does not exist.

ArchPOWER does not have mirrors so an empty mirrorlist is appropriate.
During standard install procedure pacstrap is used, copying mirrorlist.
Which in turns avoids the error when using pacstrap in the new system.